1.33+possbl. overflow fix & Warningfixes & Changed behaviour of Z-Offset-Buttons in Menupage2
Killed some warnings only Arduino 1.6.5 showed.
-> I merged the changes the original repetier did to prevent this.
Found and fixed some possible overflow within the emergency stop function introduced in 1.33.
(Printer::stepperDirection[E_AXIS] = 0;)

The Zup-ZDown-Buttons on the printer changed their behaviour if you are in Menu 2.
The problem was that if the automatic did adjust the bed to the Plus-Z (down) and you changed the origin of the z-offset you were not able to decrese the Sense-Offset. Now, if SenseOffset is present and you tune your Bed up, the Z-Offset ist not altered, but the Senseoffset is decreased. This might feel a bit strange on the display at the beginning, but turned out to be very clever.

Trick: Up-Down-Up-Down-Up-Down shifts offset from a possible Senseoffset to your static Z-Offset.