Add marker für SenseOffsetControl
Now the @ Symbol changes to ^ Symbol whenever the printer is controlling the SenseOffset. When you are at higher Z such as printing the starting line you should always see @. Whenever this is the first layer you should see ^.
This Symbol can be found in Mod-Menu (Menupage 2) only.
This Symbol is ment to debug and ment to give a more visible sign to the user that senseoffset is working.